+++ tests/bounds2_test.cpp
@@ -128,7 +128,7 @@ TEST(Bounds2Tests, BoundingSphere)
     b.BoundingSphere(center, radius);
 
     EXPECT_EQ(center, Point2f(3, 3));
-    EXPECT_FLOAT_EQ(radius, std::sqrt(2));
+    EXPECT_FLOAT_EQ(radius, std::sqrtf(2));
 }
 
 TEST(Bounds2Tests, Extent)
