+++ tests/normal3_test.cpp
@@ -154,7 +154,7 @@ TEST(Normal3Tests, Length)
 {
     Normal3f n1(3, 4, 5);
 
-    EXPECT_FLOAT_EQ(n1.Length(), std::sqrt(50));
+    EXPECT_FLOAT_EQ(n1.Length(), std::sqrtf(50));
     EXPECT_FLOAT_EQ(n1.LengthSquared(), 50);
 }
